feat(lineage): improve lineage focus UX
